cli/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "y-agent-cli"
version = "0.4.0"
description = "Personal AI agent toolkit"
requires-python = ">=3.10"
dependencies = [
    "y-agent-storage",
    "y-agent-agent",
    "click>=8.1.0",
    "prompt-toolkit>=3.0.0",
    "rich>=13.9.4",
    "httpx>=0.27.0",
    "tabulate>=0.9.0",
    "wcwidth>=0.2.13",
    "pyperclip>=1.9.0",
    "loguru>=0.7.3",
    "python-dotenv>=1.0.0",
    "python-dateutil>=2.8.0",
]

[project.scripts]
y = "yagent.command_option:cli"
y-agent = "yagent.command_option:cli"

[tool.uv.sources]
y-agent-storage = { workspace = true }
y-agent-agent = { workspace = true }

[tool.hatch.build.targets.wheel]
packages = ["src/yagent"]
